backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# LangChain and Anthropic
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# WebSocket support (included in uvicorn[standard])
websockets>=13.0

